+++ js/editor.js
@@ -2,7 +2,7 @@
 
 import {BehaviorSubject} from 'rxjs/BehaviorSubject'
 
-import {newCmdBus$ } from './lib/ndp-software/ottomann/cmdBus.js'
+import {newCmdBus$ } from './lib/ndp-software/pilota/cmdBus.js'
 
 import {currentTonality$} from './tonality.js'
 import {msPerPeriod$} from './tempo.js'
