+++ js/lib/ndp-software/pilota/resolver.js
