+++ test/pilota/dispatcher-spec.js
@@ -3,8 +3,8 @@
 import {assert} from 'chai'
 import simple from 'simple-mock'
 
-import {newObjectResolver} from '../../js/lib/ndp-software/ottomann/resolver.js'
-import {newDispatcher} from '../../js/lib/ndp-software/ottomann/dispatcher.js'
+import {newObjectResolver} from '../../js/lib/ndp-software/pilota/resolver.js'
+import {newDispatcher} from '../../js/lib/ndp-software/pilota/dispatcher.js'
 
 describe('newDispatcher', function() {
